refactor: move the forms and tests to the module directory
