+++ frontend/app/src/modules/history/management/forms/EthDepositEventForm.vue
@@ -1,43 +1,45 @@
 <script setup lang="ts">
 import type { IndependentEventData } from '@/modules/history/management/forms/form-types';
-import type { EthWithdrawalEvent, NewEthWithdrawalEventPayload } from '@/types/history/events';
-import HistoryEventAssetPriceForm from '@/components/history/events/forms/HistoryEventAssetPriceForm.vue';
+import type { EthDepositEvent, NewEthDepositEventPayload } from '@/types/history/events';
 import AmountInput from '@/components/inputs/AmountInput.vue';
 import AutoCompleteWithSearchSync from '@/components/inputs/AutoCompleteWithSearchSync.vue';
 import DateTimePicker from '@/components/inputs/DateTimePicker.vue';
+import JsonInput from '@/components/inputs/JsonInput.vue';
 import { useFormStateWatcher } from '@/composables/form';
 import { useHistoryEventsForm } from '@/composables/history/events/form';
 import { useAccountAddresses } from '@/modules/balances/blockchain/use-account-addresses';
+import HistoryEventAssetPriceForm from '@/modules/history/management/forms/HistoryEventAssetPriceForm.vue';
 import { DateFormat } from '@/types/date-format';
 import { bigNumberifyFromRef } from '@/utils/bignumbers';
 import { convertFromTimestamp, convertToTimestamp } from '@/utils/date';
 import { toMessages } from '@/utils/validation';
-import { Blockchain, HistoryEventEntryType, isValidEthAddress, Zero } from '@rotki/common';
+import { Blockchain, HistoryEventEntryType, isValidEthAddress, isValidTxHash, Zero } from '@rotki/common';
 import useVuelidate from '@vuelidate/core';
 import { helpers, required, requiredIf } from '@vuelidate/validators';
 import dayjs from 'dayjs';
 import { isEmpty } from 'es-toolkit/compat';
 
-interface EthWithdrawalEventFormProps {
-  data: IndependentEventData<EthWithdrawalEvent>;
+interface EthDepositEventFormProps {
+  data: IndependentEventData<EthDepositEvent>;
 }
 
 const stateUpdated = defineModel<boolean>('stateUpdated', { default: false, required: false });
-
-const props = defineProps<EthWithdrawalEventFormProps>();
+const props = defineProps<EthDepositEventFormProps>();
 
 const { t } = useI18n();
 
 const { data } = toRefs(props);
 
-const assetPriceForm = ref<InstanceType<typeof HistoryEventAssetPriceForm>>();
+const assetPriceForm = useTemplateRef<InstanceType<typeof HistoryEventAssetPriceForm>>('assetPriceForm');
 
+const txHash = ref<string>('');
 const eventIdentifier = ref<string>('');
 const datetime = ref<string>('');
 const amount = ref<string>('');
+const sequenceIndex = ref<string>('');
 const validatorIndex = ref<string>('');
-const withdrawalAddress = ref<string>('');
-const isExit = ref<boolean>(false);
+const depositor = ref<string>('');
+const extraData = ref<object>({});
 
 const errorMessages = ref<Record<string, string[]>>({});
 
@@ -45,20 +47,28 @@ const rules = {
   amount: {
     required: helpers.withMessage(t('transactions.events.form.amount.validation.non_empty'), required),
   },
+  depositor: {
+    isValid: helpers.withMessage(t('transactions.events.form.depositor.validation.valid'), (value: string) =>
+      isValidEthAddress(value)),
+    required: helpers.withMessage(t('transactions.events.form.depositor.validation.non_empty'), required),
+  },
   eventIdentifier: {
     required: helpers.withMessage(
       t('transactions.events.form.event_identifier.validation.non_empty'),
       requiredIf(() => get(data).type === 'edit'),
     ),
   },
+  sequenceIndex: {
+    required: helpers.withMessage(t('transactions.events.form.sequence_index.validation.non_empty'), required),
+  },
   timestamp: { externalServerValidation: () => true },
+  txHash: {
+    isValid: helpers.withMessage(t('transactions.events.form.tx_hash.validation.valid'), (value: string) =>
+      isValidTxHash(value)),
+    required: helpers.withMessage(t('transactions.events.form.tx_hash.validation.non_empty'), required),
+  },
   validatorIndex: {
     required: helpers.withMessage(t('transactions.events.form.validator_index.validation.non_empty'), required),
-  },
-  withdrawalAddress: {
-    isValid: helpers.withMessage(t('transactions.events.form.withdrawal_address.validation.valid'), (value: string) =>
-      isValidEthAddress(value)),
-    required: helpers.withMessage(t('transactions.events.form.withdrawal_address.validation.non_empty'), required),
   },
 };
 
@@ -69,10 +79,12 @@ const { getAddresses } = useAccountAddresses();
 
 const states = {
   amount,
+  depositor,
   eventIdentifier,
+  sequenceIndex,
   timestamp: datetime,
+  txHash,
   validatorIndex,
-  withdrawalAddress,
 };
 
 const v$ = useVuelidate(
@@ -83,35 +95,42 @@ const v$ = useVuelidate(
     $externalResults: errorMessages,
   },
 );
+
 useFormStateWatcher(states, stateUpdated);
 
-const withdrawalAddressSuggestions = computed(() => getAddresses(Blockchain.ETH));
+const depositorSuggestions = computed(() => getAddresses(Blockchain.ETH));
 
 function reset() {
+  set(sequenceIndex, get(data)?.nextSequenceId || '0');
+  set(txHash, '');
   set(eventIdentifier, null);
   set(datetime, convertFromTimestamp(dayjs().valueOf(), DateFormat.DateMonthYearHourMinuteSecond, true));
   set(amount, '0');
   set(validatorIndex, '');
-  set(withdrawalAddress, '');
-  set(isExit, false);
+  set(depositor, '');
+  set(extraData, {});
   set(errorMessages, {});
 
   get(assetPriceForm)?.reset();
 }
 
-function applyEditableData(entry: EthWithdrawalEvent) {
+function applyEditableData(entry: EthDepositEvent) {
+  set(sequenceIndex, entry.sequenceIndex?.toString() ?? '');
+  set(txHash, entry.txHash);
   set(eventIdentifier, entry.eventIdentifier);
   set(datetime, convertFromTimestamp(entry.timestamp, DateFormat.DateMonthYearHourMinuteSecond, true));
   set(amount, entry.amount.toFixed());
   set(validatorIndex, entry.validatorIndex.toString());
-  set(withdrawalAddress, entry.locationLabel);
-  set(isExit, entry.isExit);
+  set(depositor, entry.locationLabel);
+  set(extraData, entry.extraData || {});
 }
 
-function applyGroupHeaderData(entry: EthWithdrawalEvent) {
+function applyGroupHeaderData(entry: EthDepositEvent) {
+  set(sequenceIndex, get(data)?.nextSequenceId || '0');
   set(eventIdentifier, entry.eventIdentifier);
-  set(withdrawalAddress, entry.locationLabel ?? '');
+  set(txHash, entry.txHash);
   set(validatorIndex, entry.validatorIndex.toString());
+  set(depositor, entry.locationLabel ?? '');
   set(datetime, convertFromTimestamp(entry.timestamp, DateFormat.DateMonthYearHourMinuteSecond, true));
 }
 
@@ -126,14 +145,16 @@ async function save(): Promise<boolean> {
 
   const timestamp = convertToTimestamp(get(datetime), DateFormat.DateMonthYearHourMinuteSecond, true);
 
-  const payload: NewEthWithdrawalEventPayload = {
+  const payload: NewEthDepositEventPayload = {
     amount: get(numericAmount).isNaN() ? Zero : get(numericAmount),
-    entryType: HistoryEventEntryType.ETH_WITHDRAWAL_EVENT,
-    eventIdentifier: get(eventIdentifier),
-    isExit: get(isExit),
+    depositor: get(depositor),
+    entryType: HistoryEventEntryType.ETH_DEPOSIT_EVENT,
+    eventIdentifier: get(eventIdentifier) ?? null,
+    extraData: get(extraData) || null,
+    sequenceIndex: get(sequenceIndex) || '0',
     timestamp,
+    txHash: get(txHash),
     validatorIndex: parseInt(get(validatorIndex)),
-    withdrawalAddress: get(withdrawalAddress),
   };
 
   const eventData = get(data);
@@ -161,6 +182,7 @@ function checkPropsData() {
 }
 
 watch(data, checkPropsData);
+
 onMounted(() => {
   checkPropsData();
 });
@@ -184,7 +206,6 @@ defineExpose({
         :error-messages="toMessages(v$.timestamp)"
         @blur="v$.timestamp.$touch()"
       />
-
       <AmountInput
         v-model="validatorIndex"
         variant="outlined"
@@ -195,6 +216,16 @@ defineExpose({
         @blur="v$.validatorIndex.$touch()"
       />
     </div>
+
+    <RuiTextField
+      v-model="txHash"
+      variant="outlined"
+      color="primary"
+      data-cy="txHash"
+      :label="t('common.tx_hash')"
+      :error-messages="toMessages(v$.txHash)"
+      @blur="v$.txHash.$touch()"
+    />
 
     <RuiDivider class="mb-6 mt-2" />
 
@@ -209,29 +240,33 @@ defineExpose({
 
     <RuiDivider class="mb-6" />
 
-    <AutoCompleteWithSearchSync
-      v-model="withdrawalAddress"
-      :items="withdrawalAddressSuggestions"
-      data-cy="withdrawalAddress"
-      :label="t('transactions.events.form.withdrawal_address.label')"
-      :error-messages="toMessages(v$.withdrawalAddress)"
-      auto-select-first
-      @blur="v$.withdrawalAddress.$touch()"
-    />
+    <div class="grid md:grid-cols-2 gap-4">
+      <AutoCompleteWithSearchSync
+        v-model="depositor"
+        :items="depositorSuggestions"
+        data-cy="depositor"
+        :label="t('transactions.events.form.depositor.label')"
+        :error-messages="toMessages(v$.depositor)"
+        auto-select-first
+        @blur="v$.depositor.$touch()"
+      />
 
-    <RuiCheckbox
-      v-model="isExit"
-      color="primary"
-      data-cy="isExited"
-    >
-      {{ t('transactions.events.form.is_exit.label') }}
-    </RuiCheckbox>
+      <AmountInput
+        v-model="sequenceIndex"
+        variant="outlined"
+        integer
+        data-cy="sequenceIndex"
+        :label="t('transactions.events.form.sequence_index.label')"
+        :error-messages="toMessages(v$.sequenceIndex)"
+        @blur="v$.sequenceIndex.$touch()"
+      />
+    </div>
 
-    <RuiDivider class="mb-2" />
+    <RuiDivider class="mb-2 mt-6" />
 
     <RuiAccordions>
       <RuiAccordion
-        data-cy="eth-block-event-form__advance"
+        data-cy="eth-deposit-event-form__advance"
         header-class="py-4"
         eager
       >
@@ -247,6 +282,11 @@ defineExpose({
             :label="t('transactions.events.form.event_identifier.label')"
             :error-messages="toMessages(v$.eventIdentifier)"
             @blur="v$.eventIdentifier.$touch()"
+          />
+
+          <JsonInput
+            v-model="extraData"
+            :label="t('transactions.events.form.extra_data.label')"
           />
         </div>
       </RuiAccordion>
