+++ frontend/app/src/modules/history/management/forms/HistoryEventAssetPriceForm.vue
