+++ frontend/app/src/modules/history/management/forms/HistoryEventTypeForm.vue
